requirements.txt
numpy>=1.21
pandas>=1.3
matplotlib>=3.5
scipy>=1.7
seaborn>=0.12
pyyaml>=6.0
